requirements.txt
# Core LLM & Embeddings
openai>=1.0.0

# Vector Database
chromadb>=0.4.0

# Web Scraping
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
slowapi>=0.1.9  # Rate limiting

# UI
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
